Moving Mock utility methods into their own file.
Presumably we'll want to mock more things in the future and these
functions aren't actually specific to the WoW API.
